+++ src/File.cpp
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/File.hpp"
+#include "File.hpp"
 
 File::File(int fd) : Buffer(fd) {
     error = false;
