+++ src/Location.cpp
@@ -1,4 +1,4 @@
-#include "../include/Location.hpp"
+#include "Location.hpp"
 
 Location::Location()
 	: path(""),
