+++ src/Utils.cpp
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/Library.hpp"
+#include "Library.hpp"
 
 void setNonblock(int fd) {
 	if (fcntl(fd, F_SETFL, O_NONBLOCK) < 0) {
