+++ src/main.cpp
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/Webserv.hpp"
+#include "Webserv.hpp"
 
 int main(int ac, char *av[]) {
 	Webserv webserv;
